+++ scripts/handle_tmux_automatic_start/osx_iterm_start_tmux.sh
@@ -42,7 +42,12 @@ resize_window_to_full_screen() {
 resize_to_true_full_screen() {
 	osascript <<-EOF
 	tell application "iTerm"
+		# wait for iTerm to start
 		delay 1
+		activate
+		# short wait for iTerm to gain focus
+		delay 0.1
+		# Command + Enter for fullscreen
 		tell i term application "System Events"
 			key code 36 using {command down}
 		end tell