+++ scripts/handle_tmux_automatic_start/osx_terminal_start_tmux.sh
@@ -29,7 +29,11 @@ resize_window_to_full_screen() {
 resize_to_true_full_screen() {
 	osascript <<-EOF
 	tell application "Terminal"
+		# waiting for Terminal.app to start
 		delay 1
+		activate
+		# short wait for Terminal to gain focus
+		delay 0.1
 		tell application "System Events"
 			keystroke "f" using {control down, command down}
 		end tell